make fix

+++ src/basis_str_tool.cpp
@@ -176,7 +176,7 @@ bool BSStrTool::toUint32(const string& str, uint32& uValue, uint32 radix)
 	char *pEnd = NULL;
 	uValue = strtoul(str.c_str(), &pEnd, radix);
 	if (ERANGE == errno) return false;
-	if (pEnd - str.c_str() != str.size()) return false;
+	if ((uint32)(pEnd - str.c_str()) != str.size()) return false;
 
 	return true;
 }
@@ -189,7 +189,7 @@ bool BSStrTool::toInt32(const string& str, int32& nValue, uint32 radix)
 	char *pEnd = NULL;
 	nValue = strtol(str.c_str(), &pEnd, radix);
 	if (ERANGE == errno) return false;
-	if (pEnd - str.c_str() != str.size()) return false;
+	if ((uint32)(pEnd - str.c_str()) != str.size()) return false;
 
 	return true;
 }
@@ -202,7 +202,7 @@ bool BSStrTool::toDoulbe(const string& str, double& dValue)
 	char *pEnd = NULL;
 	dValue = strtod(str.c_str(), &pEnd);
 	if (ERANGE == errno) return false;
-	if (pEnd - str.c_str() != str.size()) return false;
+	if ((uint32)(pEnd - str.c_str()) != str.size()) return false;
 
 	return true;
 }
